+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- [crawler] gerrit add support for http basic auth.
 - [crawler] gerrit add insecure option to bypass SSL certificate verification.
 
 ### Changed